moving cache into the general purpose tools
