+++ renderer/tools/cache.js
